scripts/requirements.txt
pycryptodome==3.9.7
patch==1.16
